requirements.txt
requests
beautifulsoup4
python-dotenv
playwright
selenium
webdriver-manager
